feat (statistics) : 가게 대시보드 구현

+++ src/main/java/com/example/outsourcing/domain/statistics/dto/response/StatisticsCountResponseDto.java
@@ -9,5 +9,5 @@ public class StatisticsCountResponseDto {
 
     private final String storeName;
 
-    private final Integer orderCount;
+    private final Long orderCount;
 }
